Upgrade to AutoPlugin
This required updating all the scripted tests to use SBT version that
supports autoplugins.
The plugin, despite being named auto, will not load automatically now,
to enable opt-in behavior for multi-project builds. This is a change
from the previous version, where it was auto-loaded, hence another
modification in scripted tests to enable the plugin.